+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## [Unreleased]
 ### Changed
 - Increased the `z-index` in the `Nav` component to have it hide the content below it.
+- Improve a11y by adding an `aria-label` to the `Hamburger` button.
 
 ## [3.1.0] - 2018-08-29
 ### Added