+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [3.2.0] - 2018-09-05
 ### Added
 - The `FileField` component.
 - The `Panel.LoaderBody` subcomponent, for when the body is still being loaded.
@@ -221,9 +223,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Initial checkin.
 - The `Warning` component.
 
-[Unreleased]: https://github.com/CultureHQ/components/compare/v2.2.0...HEAD
+[Unreleased]: https://github.com/CultureHQ/components/compare/3.2.0...HEAD
+[3.2.0]: https://github.com/CultureHQ/components/compare/v3.1.0...v3.2.0
+[3.1.0]: https://github.com/CultureHQ/components/compare/v3.0.2...v3.1.0
+[3.0.2]: https://github.com/CultureHQ/components/compare/v3.0.1...v3.0.2
+[3.0.1]: https://github.com/CultureHQ/components/compare/v3.0.0...v3.0.1
+[3.0.0]: https://github.com/CultureHQ/components/compare/v2.3.0...v3.0.0
+[2.3.0]: https://github.com/CultureHQ/components/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/CultureHQ/components/compare/v2.1.1...v2.2.0
-[2.1.1]: https://github.com/CultureHQ/components/compare/v2.1.0...v2.1.1
 [2.1.1]: https://github.com/CultureHQ/components/compare/v2.1.0...v2.1.1
 [2.1.0]: https://github.com/CultureHQ/components/compare/v2.0.2...v2.1.0
 [2.0.2]: https://github.com/CultureHQ/components/compare/v2.0.1...v2.0.2